Decouple the proto generator from the data object generator
