+++ src/converters/java/io/vertx/test/codegen/converter/RecursiveItem.java
@@ -1,9 +1,11 @@
 package io.vertx.test.codegen.converter;
 
 import io.vertx.codegen.annotations.DataObject;
+import io.vertx.codegen.annotations.ProtobufGen;
 
 // Temporary Test Object, maybe will switch to test with TestDataObject
-@DataObject(generateConverter = true, protoConverter = true)
+@DataObject
+@ProtobufGen
 public class RecursiveItem {
 
   private String id;
