+++ src/converters/java/io/vertx/test/codegen/converter/User.java
@@ -1,6 +1,7 @@
 package io.vertx.test.codegen.converter;
 
 import io.vertx.codegen.annotations.DataObject;
+import io.vertx.codegen.annotations.ProtobufGen;
 import io.vertx.core.json.JsonObject;
 
 import java.time.Instant;
@@ -10,7 +11,8 @@ import java.util.Map;
 import java.util.Objects;
 
 // Temporary Test Object, maybe will switch to test with TestDataObject
-@DataObject(generateConverter = true, protoConverter = true)
+@DataObject
+@ProtobufGen
 public class User {
   private String userName;
   private Integer age;
