+++ src/main/java/io/vertx/codegen/annotations/ProtobufGen.java
@@ -1,0 +1,4 @@
+package io.vertx.codegen.annotations;
+
+public @interface ProtobufGen {
+}
